backend/requirements.txt
# Qwen3-TTS Backend Dependencies

# Core TTS package (includes torch, transformers, etc.)
qwen-tts==0.0.5

# FastAPI server
fastapi==0.128.0
uvicorn[standard]==0.40.0

# Audio processing
soundfile==0.13.1

# Authentication — JWT signing and password utilities
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
